requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
coverage==7.7.0
cycler==0.12.1
dill==0.3.9
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
etelemetry==0.3.1
factory_boy==3.3.3
Faker==37.0.2
filelock==3.18.0
fitz==0.0.1.dev2
fonttools==4.56.0
freezegun==1.5.1
frontend==0.0.3
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httplib2==0.22.0
identify==2.6.9
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
isort==6.0.1
itsdangerous==2.2.0
kiwisolver==1.4.8
looseversion==1.3.0
lxml==5.3.1
Markdown==3.7
matplotlib==3.10.1
mccabe==0.7.0
multidict==6.2.0
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nodeenv==1.9.1
numpy==2.2.4
osrm-py==0.5
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
polyline==2.0.2
pre_commit==4.1.0
propcache==0.3.0
prov==2.0.1
puremagic==1.28
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.9.0
pylint==3.3.5
pyparsing==3.2.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
reportlab==4.3.1
requests==2.32.3
scipy==1.15.2
shapely==2.0.7
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
tenacity==9.0.0
tomlkit==0.13.2
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
yarl==1.18.3
